[FEAT] imageModel 변경, experienceMainModel rating 형 변환, 이색체험 필터링 구현

+++ NanaLand/NanaLand/Experience/ExperienceMainModel.swift
@@ -17,7 +17,7 @@ struct ExperienceData: Codable {
     let firstImage: ImageList
     let title: String
     let addressTag: String
-    let ratingAvg: Int64
+    let ratingAvg: Double
     var favorite: Bool
 }
